+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -27,12 +27,12 @@ from pytorch_lightning.accelerators.tpu import TPUAccelerator
 from pytorch_lightning.plugins import (
     ApexMixedPrecisionPlugin,
     CheckpointIO,
-    DataParallelPlugin,
-    DDP2Plugin,
+    DataParallelStrategy,
+    DDP2Strategy,
     DDPFullyShardedStrategy,
-    DDPShardedPlugin,
-    DDPSpawnPlugin,
+    DDPShardedStrategy,
     DDPSpawnShardedPlugin,
+    DDPSpawnStrategy,
     DDPStrategy,
     DeepSpeedPrecisionPlugin,
     DeepSpeedStrategy,
@@ -45,7 +45,7 @@ from pytorch_lightning.plugins import (
     PrecisionPlugin,
     ShardedNativeMixedPrecisionPlugin,
     SingleDevicePlugin,
-    SingleTPUPlugin,
+    SingleTPUStrategy,
     Strategy,
     TPUBf16PrecisionPlugin,
     TPUPrecisionPlugin,
@@ -536,7 +536,7 @@ class AcceleratorConnector:
 
     @property
     def _is_sharded_training_type(self) -> bool:
-        return isinstance(self._training_type_plugin, (DDPShardedPlugin, DDPSpawnShardedPlugin))
+        return isinstance(self._training_type_plugin, (DDPShardedStrategy, DDPSpawnShardedPlugin))
 
     @property
     def _is_fully_sharded_training_type(self) -> bool:
@@ -700,7 +700,7 @@ class AcceleratorConnector:
         ):
             plugin = self.distributed_backend.training_type_plugin
         elif self.use_ddp2:
-            plugin = DDP2Plugin(parallel_devices=self.parallel_devices, cluster_environment=self.cluster_environment)
+            plugin = DDP2Strategy(parallel_devices=self.parallel_devices, cluster_environment=self.cluster_environment)
         elif self.use_ddp and self.use_deepspeed:
             plugin = DeepSpeedStrategy(
                 cluster_environment=self.select_cluster_environment(), parallel_devices=self.parallel_devices
@@ -722,7 +722,7 @@ class AcceleratorConnector:
             if use_tpu_spawn:
                 ddp_strategy_cls = TPUSpawnStrategy
             elif use_ddp_sharded:
-                ddp_strategy_cls = DDPShardedPlugin
+                ddp_strategy_cls = DDPShardedStrategy
             elif use_ddp_sharded_spawn:
                 ddp_strategy_cls = DDPSpawnShardedPlugin
             elif (
@@ -735,7 +735,7 @@ class AcceleratorConnector:
             ):
                 ddp_strategy_cls = DDPStrategy
             elif use_ddp_spawn or use_ddp_cpu_spawn:
-                ddp_strategy_cls = DDPSpawnPlugin
+                ddp_strategy_cls = DDPSpawnStrategy
             elif use_ddp_fully_sharded:
                 ddp_strategy_cls = DDPFullyShardedStrategy
             else:
@@ -745,11 +745,11 @@ class AcceleratorConnector:
                 parallel_devices=self.parallel_devices, cluster_environment=self.cluster_environment
             )
         elif self.use_dp:
-            plugin = DataParallelPlugin(parallel_devices=self.parallel_devices)
+            plugin = DataParallelStrategy(parallel_devices=self.parallel_devices)
         elif self.use_horovod:
             plugin = HorovodStrategy(parallel_devices=self.parallel_devices)
         elif self.use_tpu and isinstance(self.tpu_cores, list):
-            plugin = SingleTPUPlugin(self.tpu_id)
+            plugin = SingleTPUStrategy(self.tpu_id)
         elif self.use_ipu:
             plugin = IPUStrategy(parallel_devices=self.parallel_devices)
         else:
@@ -1026,8 +1026,8 @@ class AcceleratorConnector:
                     f"The `TPUAccelerator` can only be used with a `TPUPrecisionPlugin`,"
                     f" found: {self.training_type_plugin.precision_plugin}."
                 )
-            if not isinstance(self.training_type_plugin, (SingleTPUPlugin, TPUSpawnStrategy)):
+            if not isinstance(self.training_type_plugin, (SingleTPUStrategy, TPUSpawnStrategy)):
                 raise ValueError(
-                    "The `TPUAccelerator` can only be used with a `SingleTPUPlugin` or `TPUSpawnStrategy`,"
+                    "The `TPUAccelerator` can only be used with a `SingleTPUStrategy` or `TPUSpawnStrategy`,"
                     f" found {self.training_type_plugin}."
                 )