+++ tests/plugins/test_plugins_registry.py
@@ -17,9 +17,9 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.plugins import (
     CheckpointIO,
     DDPFullyShardedStrategy,
-    DDPShardedPlugin,
-    DDPSpawnPlugin,
+    DDPShardedStrategy,
     DDPSpawnShardedPlugin,
+    DDPSpawnStrategy,
     DDPStrategy,
     DeepSpeedStrategy,
     TPUSpawnStrategy,
@@ -110,9 +110,9 @@ def test_fsdp_strategys_registry(tmpdir):
     "plugin_name, plugin",
     [
         ("ddp_find_unused_parameters_false", DDPStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnPlugin),
+        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
         ("ddp_sharded_spawn_find_unused_parameters_false", DDPSpawnShardedPlugin),
-        ("ddp_sharded_find_unused_parameters_false", DDPShardedPlugin),
+        ("ddp_sharded_find_unused_parameters_false", DDPShardedStrategy),
     ],
 )
 def test_ddp_find_unused_parameters_training_type_plugins_registry(tmpdir, plugin_name, plugin):